Build: Update gradle distribution to 7.5
Command: ./gradlew wrapper --gradle-version=7.5 --distribution-type=all

Release Notes: https://docs.gradle.org/7.5/release-notes.html
Distribution: https://services.gradle.org/distributions/
gradle-7.5.1-all.zip

------------------------------------------------------------------------

Note, although the 'distributionUrl' was already pointing to
'gradle-7.5-all.zip', the 'jar' and 'gradlew' files weren't up-to-date.
This was probably because when the previous update happened, the update
was either done manually by just updating the version on that url, or
the command was run once, and not twice, which is when it actually
works and updates all relevant Gradle related files.