Build: Set org gradle jvmargs to fix an out of memory error on ci
While running the 'Publish to S3' step on CI the build fails with the
below exception:

"Exception in thread "DefaultDispatcher-worker-2"
java.lang.OutOfMemoryError: Metaspace
Daemon will be stopped at the end of the build after running out of
JVM memory"

This error occurs when the Metaspace region of the JVM's memory on CI is
exhausted. As such, the default Metaspace size may be too small for the
CI build process. Increasing the Metaspace size by adding the JVM
arguments will hopefully resolve this problem.